\docType{package}
\name{Interpol.T-package}
\alias{Interpol.T}
\alias{Interpol.T-package}
\title{Hourly interpolation of multiple temperature daily series}
\description{
  Hourly interpolation of multiple temperature daily series
}
\details{
  \tabular{ll}{ Package: \tab Interpol.T\cr Type: \tab
  Package\cr Version: \tab 2.1\cr Date: \tab 2012-07-30\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr
  Depends: \tab R(>=2.12)\cr }
}
\author{
  Emanuele Eccel & Emanuele Cordano
  \email{emanuele.eccel@fmach.it}
}
\keyword{hourly}
\keyword{interpolation}
\keyword{package,}
\keyword{temperature,}

